// src/libs/identity/enums/error-codes.enum.ts
export enum IdentityErrorCodes {
    USERNAME_NOT_FOUND = 'USER_NOT_FOUND',
    USERNAME_IS_NOT_VALID = 'INVALID_USERNAME',
    PASSWORD_IS_WRONG = 'WRONG_PASSWORD',
    TOKEN_IS_EXPIRED = 'TOKEN_EXPIRED',
    TOKEN_IS_NOT_VALID = 'INVALID_TOKEN',
    PASSWORD_IS_INVALID = 'INVALID_PASSWORD',
    DUPLICATE_USERNAME = 'DUPLICATE_USERNAME',
    VALIDATION_ERROR = 'VALIDATION_ERROR',
}